+++ src/Interfaces/subscription/SubscriptionTier.ts
@@ -1,7 +1,7 @@
 export type SubscriptionTierOrder = 1 | 2 | 3 | 4;
 export type SubscriptionTierAge = 0 | 1 | 2 | 3;
 
-export interface SubscriptionTier {
+export default interface SubscriptionTier {
     id?: string;
     subscriptionId?: string;
     order?: SubscriptionTierOrder;