+++ src/Models/subscription/SubscriptionAddon.ts
@@ -1,0 +1,7 @@
+import SubscriptionAddonInterface from '../../Interfaces/subscription/SubscriptionAddon';
+
+export default class SubscriptionAddon {
+    constructor(parameters: SubscriptionAddonInterface) {
+        Object.assign(this, parameters);
+    }
+}