+++ src/Models/subscription/Transformers/SubscriptionAddonPricingPlanTransformer.ts
@@ -1,0 +1,20 @@
+import BaseTransformer from '../../../BaseTransformer';
+import SubscriptionAddonPricingPlanInterface from '../../../Interfaces/subscription/SubscriptionAddonPricingPlan';
+import SubscriptionAddonPricingPlan from '../SubscriptionAddonPricingPlan';
+
+export default class SubscriptionAddonPricingPlanTransformer extends BaseTransformer {
+    mapData(subscriptionAddonPricingPlan: SubscriptionAddonPricingPlanInterface): SubscriptionAddonPricingPlan {
+        return new SubscriptionAddonPricingPlan({
+            id: subscriptionAddonPricingPlan.id,
+            createdAt: subscriptionAddonPricingPlan.createdAt,
+            updatedAt: subscriptionAddonPricingPlan.updatedAt,
+            name: subscriptionAddonPricingPlan.name,
+            stripeId: subscriptionAddonPricingPlan.stripeId,
+            cost: subscriptionAddonPricingPlan.cost,
+            intervalLength: subscriptionAddonPricingPlan.intervalLength,
+            intervalUnit: subscriptionAddonPricingPlan.intervalUnit,
+            archivedAt: subscriptionAddonPricingPlan.archivedAt,
+            subscriptionAddonId: subscriptionAddonPricingPlan.subscriptionAddonId,
+        });
+    }
+}